# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "A-Share Stock Backtesting & Quant Platform API",
    "version": "1.0.0",
    "description": "A-share 量化平台后端 REST API。\n\n主要业务领域：\n- **认证 & 用户**：登录/注册/JWT 刷新\n- **回测 & 策略**：单股回测、组合级策略、参数优化\n- **因子**：因子库管理、IC 报告、相关性矩阵\n- **模拟交易**：持仓、订单、风控守卫\n- **量化引擎**：StrategyEngine / SignalEngine / BacktestEngine 等 5 大 facade\n- **数据 & 行情**：股票元信息、K 线、实时行情、龙虎榜、北向资金\n- **AI**：TradingAgents 深度分析、KOL 观点、公告 NLP、市场情绪\n- **推送 & 告警**：飞书 / 邮件 / 微信 / 短信 4 channel\n- **运维 & 监控**：env 校验、CI 状态、健康度检查\n\n所有 endpoint 默认返回 `{ success: boolean, data?: any, message?: string }` envelope。\n需要鉴权的 endpoint 需在 Header 携带 `Authorization: Bearer <JWT>`。",
    "contact": {
      "name": "QuantX A-Share Alpha Refactor"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development (本地)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在 Authorization Header 携带 `Bearer <JWT>`。Token 通过 `/api/auth/login` 获取。"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "业务数据（结构因 endpoint 而异）"
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "用户 Users"
        ],
        "summary": "获取所有用户列表 (管理员)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无管理员权限"
          }
        }
      },
      "post": {
        "tags": [
          "用户 Users"
        ],
        "summary": "创建新用户 (管理员)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "email",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "user"
                    ]
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无管理员权限"
          }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": [
          "用户 Users"
        ],
        "summary": "更新用户信息 (管理员)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "user"
                    ]
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无管理员权限"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      },
      "delete": {
        "tags": [
          "用户 Users"
        ],
        "summary": "删除用户 (管理员)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无管理员权限"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
    "/api/users/{id}/password": {
      "put": {
        "tags": [
          "用户 Users"
        ],
        "summary": "修改用户密码 (管理员)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "newPassword"
                ],
                "properties": {
                  "newPassword": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无管理员权限"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
...
  }
}